+++ src/components/calcite-notice/calcite-notice.tsx
@@ -58,7 +58,7 @@ export class CalciteNotice {
   /** specify the scale of the notice, defaults to m */
   @Prop({ mutable: true, reflect: true }) scale: "s" | "m" | "l" = "m";
 
-  /** specify the scale of the button, defaults to m */
+  /** specify the width of the notice, defaults to m */
   @Prop({ mutable: true, reflect: true }) width: "auto" | "half" | "full" =
     "auto";
 
@@ -88,8 +88,8 @@ export class CalciteNotice {
 
   componentDidLoad() {
     this.noticeLinkEl = this.el.querySelectorAll(
-      "calcite-button"
-    )[0] as HTMLCalciteButtonElement;
+      "calcite-link"
+    )[0] as HTMLCalciteLinkElement;
   }
 
   render() {
@@ -173,7 +173,7 @@ export class CalciteNotice {
   private closeButton?: HTMLElement;
 
   /** the notice link child element  */
-  private noticeLinkEl?: HTMLCalciteButtonElement;
+  private noticeLinkEl?: HTMLCalciteLinkElement;
 
   private iconDefaults = {
     green: "checkCircle",